requirements.txt
aiogram==3.4.1
APScheduler==3.10.4
python-dotenv==1.0.0
SQLAlchemy==2.0.27
aiosqlite==0.19.0
aiofiles==23.2.1
pytz==2023.3
